Artikli i povezivanje izbornika Artikli i Dobavljaci sa glavnim izbornikom

+++ CSHARP/Ucenje/UcenjeCS/ZavrsniRadDamirB/Model/Artikl.cs
@@ -6,6 +6,6 @@
 
         public string? NazivArtikla { get; set; }
 
-        public int? Cijena { get; set; }
+        public float? Cijena { get; set; }
     }
 }
